+++ R/reverse.R
@@ -8,7 +8,7 @@ check_wrt_arrayish <- function(args_flat, is_wrt_flat) {
         ))
       }
 
-      if (!inherits(dtype_abstract(args_flat[[i]]), "FloatType")) {
+      if (!is_dtype_float(dtype_abstract(args_flat[[i]]))) {
         cli_abort(c(
           "Can only compute gradient with respect to float arrays.",
           x = "Got {repr(dtype_abstract(args_flat[[i]]))}"